# 02-express-api/src/configs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IWA-MEAN - Express API API",
    "version": "1.0.0",
    "description": "This is the API for IWAExpress (Insecure Web App) Pharmacy Direct an insecure NodeJS/Express web application documented with Swagger",
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "contact": {
      "name": "Fortify by Opentext",
      "url": "https://www.microfocus.com/en-us/cyberres/application-security"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/site/status": {
      "get": {
        "tags": [
          "Site"
        ],
        "summary": "Get the site status",
        "description": "Get the site message of the day",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/site/register-user": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Register a new user",
        "description": "Register a new user with the site",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "409": {
            "description": "User Already Exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/registerUser"
              }
            }
          }
        }
      }
    },
    "/api/site/subscribe-user": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Subscribe a new user",
        "description": "Subscribe a new user to the newsletter",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "409": {
            "description": "User Already Exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/subscribeUser"
              }
            }
          }
        }
      }
    },
    "/api/site/sign-in": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Sign-in",
        "description": "Sign in as an authorised user",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/jwtJson"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/jwtJson"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signInUser"
              }
            }
          }
        }
      }
    },
    "/api/site/sign-out": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Sign-out",
        "description": "Sign out an authorised user",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signOutUser"
              }
            }
          }
        }
      }
    },
    "/api/site/refresh-token": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Refresh token",
        "description": "Refresh users JwtAuthHandler access token",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/refreshUser"
              }
            }
          }
        }
      }
    },
    "/api/site/backup-newsletter-db": {
      "post": {
        "tags": [
          "Site"
        ],
        "summary": "Backup the newsletter database",
        "description": "Compress and backup the newsletter database to the specified file",
        "parameters": [
          {
            "name": "file_path",
            "in": "query",
            "description": "The file to backup the database to. Cannot be empty.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "registerUser": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "example": "User"
          },
          "last_name": {
            "type": "string",
            "example": "One"
          },
          "email": {
            "type": "string",
            "example": "user1@localhost.com"
          },
          "phone_number": {
            "type": "string",
            "example": "1234567890"
          },
          "password": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email",
          "phone_number",
          "password"
        ],
        "xml": {
          "name": "registerUser"
        }
      },
      "subscribeUser": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "example": "User"
          },
          "last_name": {
            "type": "string",
            "example": "One"
          },
          "email": {
            "type": "string",
            "example": "user1@localhost.com"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email"
        ],
        "xml": {
          "name": "subscribeUser"
        }
      },
      "signInUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "string"
          },
          "password": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "signInUser"
        }
      },
      "signOutUser": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "accessToken"
        ],
        "xml": {
          "name": "signOutUser"
        }
      },
      "refreshUser": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "xml": {
          "name": "refreshUser"
        }
      },
...
      "success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "success"
        }
      },
      "failure": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failure"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "failure"
        }
      },
      "jwtJson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "string"
          },
          "email": {
            "type": "string",
            "example": "string"
          },
          "accessToken": {
            "type": "string",
            "example": "string"
          },
          "refreshToken": {
            "type": "string",
            "example": "string"
          },
          "tokenExpiration": {
            "type": "number",
            "example": 1000000
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          }
        },
        "required": [
          "id",
          "email",
          "accessToken",
          "refreshToken",
          "tokenExpiration",
          "tokenType"
        ],
        "xml": {
          "name": "jwtJson"
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic Authorization header."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT Authorization header."
      }
    }
  },
  "security": [
    {
      "basicAuth": [],
      "bearerAuth": []
    }
  ]
}